move `event_content_is_seer_eligible` to ingest seer module
